+++ templateless/components/__init__.py
@@ -5,3 +5,6 @@ from .otp import Otp  # noqa
 from .socials import Socials, Item as SocialItem, Service  # noqa
 from .text import Text  # noqa
 from .view_in_browser import ViewInBrowser  # noqa
+from .store_badges import StoreBadges, Item as StoreBadgeItem, StoreBadge  # noqa
+from .qr_code import QrCode, Cryptocurrency  # noqa
+from .signature import Signature, Font as SignatureFont  # noqa