+++ templateless/components/image.py
@@ -2,7 +2,7 @@ from ..common import ComponentId, Component
 
 
 class Image(Component):
-    def __init__(self, src, alt=None, width=None, height=None, url=None):
+    def __init__(self, src, url=None, width=None, height=None, alt=None):
         self.id = ComponentId.IMAGE
         self.src = src
         self.alt = alt
@@ -10,8 +10,8 @@ class Image(Component):
         self.height = height
         self.url = url
 
-    def to_dict(self):
-        return {
+    def to_dict(self) -> dict:
+        attributes = {
             "id": self.id.value,
             "src": self.src,
             "alt": self.alt,
@@ -19,3 +19,4 @@ class Image(Component):
             "height": self.height,
             "url": self.url,
         }
+        return {key: value for key, value in attributes.items() if value is not None}