+++ templateless/components/link.py
@@ -7,7 +7,7 @@ class Link(Component):
         self.text = text
         self.url = url
 
-    def to_dict(self):
+    def to_dict(self) -> dict:
         return {
             "id": self.id.value,
             "text": self.text,