+++ CHANGELOG.md
@@ -1,16 +1,29 @@
 # CHANGELOG
 
-**v0.1.0-alpha.3:**
+## v0.1.0-alpha.4:
+
+### New Features
+
+- New social icons: `Service.MASTODON` and `Service.RSS`
+- New store badges component
+- New QR code component
+- New signature component
+
+### Bug fixes
+
+- Image param order has been fixed (`url` and `alt` params swapped)
+
+## v0.1.0-alpha.3:
 - `README.md`: notice about test mode
 - Support for test mode logging
 
-**v0.1.0-alpha.2:**
+## v0.1.0-alpha.2:
 - `README.md` cleanup (listing of components)
 - `Image` component now requires only `src`; the other params are optional
 - `ViewInBrowser` component has changed: text is optional
 - New example in `examples/confirm_email.py`
 
-**v0.1.0-alpha.1:**
+## v0.1.0-alpha.1:
 - Introduced `CHANGELOG.md`
 - Introduced new services as social icons:
   - `Service.PHONE` (converts into a link with `tel:` prefix)
@@ -25,5 +38,5 @@
   - `Service.THREADS`
   - `Service.TELEGRAM`
 
-**v0.1.0-alpha.0:**
+## v0.1.0-alpha.0:
 - Initial implementation